Migrate Ordering.API to WebApplicationBuilder
