+++ src/Tests/Services/Application.FunctionalTests/Services/Catalog/CatalogScenariosBase.cs
@@ -14,7 +14,7 @@ public class CatalogScenariosBase
             {
                 cb.AddJsonFile("Services/Catalog/appsettings.json", optional: false)
                 .AddEnvironmentVariables();
-            }).UseStartup<Startup>();
+            });
 
         var testServer = new TestServer(hostBuilder);
